+++ erfinv.h
@@ -14,7 +14,7 @@
 
 #include <math.h>
 
-static double erfinv(double x) {
+static float erfinv(float x) {
     assert(0.0 <= x && x < 1.0);
     if (x <= 0.9)
 	return  x * (1.1117646622106996 + x *
@@ -23,7 +23,7 @@ static double erfinv(double x) {
         return  x * (2257.764033265176 + x *
 	          (x * 883.953153922680 - 2444.745829241307)) -
 	            694.999708131834343;
-    return 1.786158859307387 + 0.004 * exp((x - 0.99) * 600);
+    return 1.786158859307387 + 0.004 * expf((x - 0.99) * 600);
 }
 
 #endif
