+++ gaussian-erfinv.h
@@ -10,9 +10,9 @@
 #include <limits.h>
 #include "erfinv.h"
 
-static double gaussian(double sd) {
+static float gaussian(float sd) {
     long r = random();
-    double y = erfinv((r & ~1) / (double)(RAND_MAX - 1));
+    float y = erfinv((r & ~1) / (float)(RAND_MAX - 1));
     if (r & 1)
 	return  -y * sd;
     return y * sd;
